Moving main.py to root
